basic component system, renderer, spriterenderer have been implmented

+++ KTEngine/Scene/DefaultScene.cpp
@@ -102,7 +102,7 @@ void DefaultScene::update(float dt)
 void DefaultScene::imgui()
 {
 	ImGui::Begin("This is a Default Scene!");
-	ImGui::Text("Application average %.3f ms/frame (%.1f FPS)", (float)Window::getFrameRate() * 1000.0f, 1.0f / Window::getFrameRate());
+	ImGui::Text("Application average %.3f ms/frame (%.1f FPS)", (double)Window::getFrameRate() * (double)1000.0f, 1.0f / Window::getFrameRate());
 	ImGui::Text("Lost Tower big step");
 	ImGui::NewLine();
